[GTK] Crash on inputting booking field on IRCTC website in ScriptMessageClientGtk::didPostMessage https://bugs.webkit.org/show_bug.cgi?id=280150
Reviewed by NOBODY (OOPS!).

Epiphany is failing to unregister a script message handler, causing it
to be emitted after its corresponding WebKitWebView and
WebKitUserContentManager have been destroyed. There are basically three
ways we could handle this: (a) detect it and ignore it, (b) detect it
and emit a warning or critical, and (c) proactively prevent it by
storing a HashMap containing all message handlers and unregistering any
that still exist when the WebKitUserContentManager is destroyed.

Maybe (c) would be best, but I'm lazy, so I chose option (b).

I spent a little while trying to create an API test for this, but it
doesn't seem simple.

* Source/WebKit/UIProcess/API/glib/WebKitUserContentManager.cpp: