[GTK] Crash when closing the Print dialog with Escape keyboard shortcut https://bugs.webkit.org/show_bug.cgi?id=289271
Reviewed by NOBODY (OOPS!).

If you click on the Cancel button, the application is expected to
destroy the dialog, but the Escape key is a shortcut for "destroy" and
GTK will destroy the dialog itself. Attempting to do anything with the
dialog after it has returned the delete event response is a use after
free.

* Source/WebKit/UIProcess/API/gtk/WebKitPrintOperation.cpp:
(webkitPrintOperationRunDialog):