[GTK] Color chooser opens in a dialog outside the window https://bugs.webkit.org/show_bug.cgi?id=259613
Reviewed by NOBODY (OOPS!).

GtkColorChooserDialog is deprecated, so use GtkColorDialog instead if it
is available. GtkColorDialog is notably a modal window, so this fixes
the color dialog opening in an entirely separate window from the rest of
the application.

The downside is that it is modal for the window rather than modal for
the web view, so you cannot interact with other tabs if any tab is
displaying a color picker. Perhaps this is not ideal. I think the nicest
approach would be to embed the dialog inside the WebKitWebViewBase, but
I'm not sure how we would possibly do that since it's no longer a
GtkWidget.

* Source/WebCore/platform/gtk/GUniquePtrGtk.h:
* Source/WebKit/UIProcess/gtk/WebColorPickerGtk.cpp:
(WebKit::WebColorPickerGtk::WebColorPickerGtk):
(WebKit::WebColorPickerGtk::~WebColorPickerGtk):
(WebKit::WebColorPickerGtk::endPicker):
(WebKit::WebColorPickerGtk::colorDialogChooseCallback):
(WebKit::WebColorPickerGtk::showColorPicker):
* Source/WebKit/UIProcess/gtk/WebColorPickerGtk.h: