[WPE][GTK] Add API to configure proxy autoconfig https://bugs.webkit.org/show_bug.cgi?id=291316
Reviewed by NOBODY (OOPS!).

WebKit already supports proxy autoconfig, but only if configured via
system proxy settings. It also supports a custom proxy autoconfig
setting (not matching the system setting) only for use by WebDriver.
Let's add this support to the API as well, so applications can provide a
proxy autoconfig URL setting in their proxy configuration dialogs. Most
GNOME apps probably won't ever use this because GNOME apps should use
system proxy settings, but some applications like Evolution want to
allow the user the flexibility to use proxy autoconfig only for the app
and not for the rest of the system.

I had a few API design decisions to make here. First, I wanted to
reflect that proxy autoconfig may not actually be supported by the
system -- the API currently depends on glib-networking, but this may
change in the future -- so I added a non-method function that tests
whether proxy autoconfig is supported, and made the new constructor
failable.

My biggest question was whether to add a new WebKitNetworkProxyMode,
WEBKIT_NETWORK_PROXY_MODE_AUTOCONFIG. I really wanted to, to match
the SoupNetworkProxySettingsMode enum, but ultimately this made things
too complicated. Problem is the WebKitNetworkProxySettings object keeps
track of its own mode, but also the user is expected to separately pass
the mode to webkit_network_session_set_proxy_settings(). We could check
and emit a critical if the application passes the wrong
WebKitNetworkProxySettings object with the wrong WebKitNetworkProxyMode,
but it's nicer and easier to dodge this problem altogether by not
exposing any new mode. WEBKIT_NETWORK_PROXY_MODE_CUSTOM will just mean
"check the WebKitNetworkProxySettings," same as it did before. There are
just two different types of WebKitNetworkProxySettings objects now: the
original one that wraps GSimpleProxyResolver, and the new one that just
contains the PAC URL.