Clang 21 complains about implicit casting between char32_t and char16_t throughout the codebase https://bugs.webkit.org/show_bug.cgi?id=299689
Reviewed by NOBODY (OOPS!).

This pull request is a WIP draft. It is incomplete and not ready for review.

This prepares WebKit's cross-platform code for Clang 21's
-Wcharacter-conversion. In practice, this warning is emitted whenever we
implicitly convert from char32_t to char16_t, or vice-versa. In general,
the solution is to add a cast.

Casting to char16_t to char32_t is safe if you're certain that none of
the code units are surrogates, so let's add a cast that asserts this is
true.

Converting from char32_t to char16_t is safe if you're certain that the
character is in the Basic Multilingual Plane, so let's add a cast that
asserts this is true.

Since char32_t is not compatible with ICU's UChar32, and we want to use
char32_t to represent UTF-32, I've added variants of various ICU
conversion macros that return char32_t instead of UChar32. They also
receive inputs as spans, use char8_t instead of UChar8, and notably
avoid triggering -Wunsafe-buffer-usage.

In a few cases, I change behavior where the original code looks wrong
and seems easy to fix, e.g. isStrWhiteSpace in ParseInt.h, or
URLParser::isForbiddenHostCodePoint.

EncodingTables.h uses templates throughout the file and I'm not sure how
to fix it, so I decided not to worry about it and just suppressed the
warning. I made the same decision for all of TestWebKitAPI to avoid
requiring changes to GoogleTest.

Lastly, I have added a deprecated/broken cast for use in cases that look
suspicious, where I suspect that the assertion might be hit. These will
require follow-up investigation.